Backed out 2 changesets (bug 1912695) for causing bustages on WasmBCMemory.cpp CLOSED TREE
Backed out changeset 7961243343b2 (bug 1912695)
Backed out changeset 4fa0b44a5701 (bug 1912695)

+++ js/src/wasm/WasmIonCompile.cpp
@@ -1361,7 +1361,7 @@ class FunctionCompiler {
   // the offset rather than vice versa is that a small offset can be ignored
   // by both explicit bounds checking and bounds check elimination.
   void foldConstantPointer(MemoryAccessDesc* access, MDefinition** base) {
-    uint64_t offsetGuardLimit = GetMaxOffsetGuardLimit(
+    uint32_t offsetGuardLimit = GetMaxOffsetGuardLimit(
         codeMeta_.hugeMemoryEnabled(access->memoryIndex()));
 
     if ((*base)->isConstant()) {
@@ -1373,8 +1373,9 @@ class FunctionCompiler {
       }
 
       uint64_t offset = access->offset64();
+
       if (offset < offsetGuardLimit && basePtr < offsetGuardLimit - offset) {
-        offset += basePtr;
+        offset += uint32_t(basePtr);
         access->setOffset32(uint32_t(offset));
         *base = isMem64(access->memoryIndex()) ? constantI64(int64_t(0))
                                                : constantI32(0);
@@ -1386,7 +1387,7 @@ class FunctionCompiler {
   // be checked, compute the effective address, trapping on overflow.
   void maybeComputeEffectiveAddress(MemoryAccessDesc* access,
                                     MDefinition** base, bool mustAddOffset) {
-    uint64_t offsetGuardLimit = GetMaxOffsetGuardLimit(
+    uint32_t offsetGuardLimit = GetMaxOffsetGuardLimit(
         codeMeta_.hugeMemoryEnabled(access->memoryIndex()));
 
     if (access->offset64() >= offsetGuardLimit ||
@@ -1475,8 +1476,8 @@ class FunctionCompiler {
     MOZ_ASSERT(!inDeadCode());
     MOZ_ASSERT(!codeMeta_.isAsmJS());
 
-    // Attempt to fold a constant base pointer into the offset so as to simplify
-    // the addressing expression. This may update *base.
+    // Attempt to fold an offset into a constant base pointer so as to simplify
+    // the addressing expression.  This may update *base.
     foldConstantPointer(access, base);
 
     // Determine whether an alignment check is needed and whether the offset
@@ -1540,8 +1541,7 @@ class FunctionCompiler {
     return codeMeta_.hugeMemoryEnabled(memoryIndex);
   }
 
-  // Add the offset into the pointer to yield the EA; trap on overflow. Clears
-  // the offset on the memory access as a result.
+  // Add the offset into the pointer to yield the EA; trap on overflow.
   MDefinition* computeEffectiveAddress(MDefinition* base,
                                        MemoryAccessDesc* access) {
     if (inDeadCode()) {
